requirements.txt
openai-whisper
pydub
ffmpeg-python
opencv-python
torch>=1.9.0
torchvision>=0.10.0
pytorchvideo
numpy
fvcore
av